extract out common ontology loader elements
